requirements.txt
Flask==2.0.1
PyYaml==5.4.1
gunicorn==20.1.0
Flask-RESTful==0.3.9